app/requirements.txt
opencv-python-headless 
torch 
torchvision 
ultralytics 
requests
python-dotenv
python-telegram-bot